+++ edition/voxel_tool_lod_terrain.cpp
@@ -14,7 +14,7 @@ bool VoxelToolLodTerrain::is_area_editable(const Rect3i &box) const {
 	return _map->is_area_fully_loaded(box.padded(1));
 }
 
-int VoxelToolLodTerrain::_get_voxel(Vector3i pos) {
+uint64_t VoxelToolLodTerrain::_get_voxel(Vector3i pos) {
 	ERR_FAIL_COND_V(_terrain == nullptr, 0);
 	return _map->get_voxel(pos, _channel);
 }
@@ -24,7 +24,7 @@ float VoxelToolLodTerrain::_get_voxel_f(Vector3i pos) {
 	return _map->get_voxel_f(pos, _channel);
 }
 
-void VoxelToolLodTerrain::_set_voxel(Vector3i pos, int v) {
+void VoxelToolLodTerrain::_set_voxel(Vector3i pos, uint64_t v) {
 	ERR_FAIL_COND(_terrain == nullptr);
 	_map->set_voxel(v, pos, _channel);
 }
